requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Admin==1.6.1
Flask-Limiter==3.5.0
Flask-WTF==1.2.1
python-dotenv==1.0.0
Pillow==10.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.20
pytest==7.4.0 